Vision_Picture_RAG/requirements.txt
SQLAlchemy
psycopg2-binary
pgvector[sqlalchemy]
# Embedding / LLM dependencies
sentence-transformers
# Qwen2-VL wymaga nowszego transformers
transformers>=4.40.0
accelerate
# PyTorch z obsługą CUDA (dla GPU)
torch>=2.0.0
torchvision>=0.15.0
# Jeśli masz GPU NVIDIA, zainstaluj: pip install torch --index-url https://download.pytorch.org/whl/cu118
Pillow
# Dodatkowe przyspieszenie
bitsandbytes  # Optymalizacja modeli na GPU
qwen-vl-utils>=0.0.8
timm>=0.9.8
einops